Refactor extract command tests to package main(command_test)
